Fix permissions of etc/symfony_app.secret
It needs to be readable by the webserver.

DOMjudge/domjudge/Makefile installs it using $(INSTALL_WEBSITE), i.e.
with owner $DOMJUDGE_USER and group $WEBSERVER_GROUP (same as
dbpasswords.secret), but the ownership is lost when the Dockerfile
copies the domserver directory.

Set the owner and group to www-data, consistently with
dbpasswords.secrets.