Manually deploy release dockers to DockerHUB
This is only triggered from master, in GitLab there is a button to
trigger the release.

The current docker:latest image does not have curl, because of this the
`-s` option is removed which yields the problem that when the server
does not have the release but returns a page we will try to unpack that
page. As thats a failure its fine for now. Additionally this assumes
that if you try to build locally you need wget.