Manual on the created debs
Part of the jobs to automate new releases. This can only be triggered from master, when the new debs are created these jobs can be used to see if there are installation issues on debian like distros.

Implemented tests:
- Ubuntu LTS
- Ubuntu latest
- Debian stable
- Debian oldstable

In GitLab there is a button to trigger the pipeline.

Also renamed the stage to the default test.