Jobs to automate new releases
This is only triggered from master, these jobs are triggered from the
domjudge.org server.

Implemented tests:
- Test the new tarball
- Do a test install
- Deploy new docker containers

in GitLab there is a button to
trigger the release.

The current docker:latest image does not have curl, because of this the
`-s` option is removed which yields the problem that when the server
does not have the release but returns a page we will try to unpack that
page. As thats a failure its fine for now. Additionally this assumes
that if you try to build locally you need wget.

Job is added to release a new DOMjudge version

The release script is stored in domjudge-scripts. For instructions on
which steps to take one should check the howto in:

- https://github.com/DOMjudge/domjudge-scripts/blob/master/new_release_howto.txt

The build will be signed after building and tested and if an
installation works will be uploaded to the DOMjudge.org server.

The GPG key can be found in the gitlab settings in the variables
setting. The SSH key for the rsync is in the same location.

The job is added in another file to keep it readable. Jobs can be
collapsed.

[DELETE] fix for having the pipeline pass
[DELETE] Hardcode the version for the PR

go