Job is added to build a new DOMjudge tarball
The release script is stored in domjudge-scripts. For instructions on
which steps to take one should check the howto in:

- https://github.com/DOMjudge/domjudge-scripts/blob/master/new_release_howto.txt

The build will not signed after building but can be used to already run
some tests with.

The job is added in another file to keep it readable. Jobs can be
collapsed.

The final tarball will be build based on the signed tag on the
domjudge.org.