Add HEALTHCHECK to domserver Docker container
When running the Docker container for domserver, it is possible that one
of nginx or php-fpm crashes/stops working. When this happens from
outside the container the effects can be quite mysterious. Adding an
health check allows the user to diagnose more quickly what happened.

Running `docker ps`, in the column `STATUS` it is shown: `Up X minutes
(healthy)` or `Up X minutes (unhealthy)` depending on the exit code of
the health check script. To further diagnose the problem one can run
`docker inspect domserver` (optionally piped to `jq '.[0].State.Health'`)
and this will report diagnostic information like:

```json
{
  "Status": "healthy",
  "FailingStreak": 0,
  "Log": [
    {
      "Start": "2021-12-19T19:22:53.555071655+01:00",
      "End": "2021-12-19T19:22:54.012109556+01:00",
      "ExitCode": 0,
      "Output": "nginx ok | php ok | http ok"
    },
    ...
  ]
}
```

This health check probes supervisord for the status of the two services
(nginx and php-fpm), as well as tries to reach the API with an HTTP
request.

This commit also fixes the deprecated `MAINTAINER` labels in the various
Dockerfiles.