Only measure coverage on OS PHP
Alternative is to remove the pecl calls and use `apt install php*-pecl`
as this is now provided. We discussed and measuring the coverage
multiple times is not needed.

As a contributor I think we also prefer to not measure the coverage
already when testing with another PHP version but only in the CI.