Switch from CI_PRE_CLONE_SCRIPT to FF_DISABLE_UMASK_FOR_DOCKER_EXECUTOR
CI_PRE_CLONE_SCRIPT is deprecated and will be removed in GitLab 16.0
(see [1]).

We use it to disable "umask 0000", which can also be done using
FF_DISABLE_UMASK_FOR_DOCKER_EXECUTOR (see [2]), so switch to that.

Also add a regression test.

(When I was implementing the original fix using CI_PRE_CLONE_SCRIPT in
#137, I did also try FF_DISABLE_UMASK_FOR_DOCKER_EXECUTOR, but it
didn't seem to work. It works fine now; I think the issue at the time
was due to caching, see [3].)

[1]: https://docs.gitlab.com/ee/update/deprecations?removal_milestone=16.0&breaking_only=true#deprecation-and-planned-removal-for-ci_pre_clone_script-variable-on-gitlab-saas
[2]: https://www.gitlab.com/gitlab-org/gitlab-runner/-/issues/1736#note_1370906098
[3]: https://www.gitlab.com/gitlab-org/gitlab/-/issues/300715