Repair the release
The repo was not properly quoted and the github.ref has another format with direct push.
The push should be done without the tag after the image.

Push resulting image to our DOMjudge GitHub container registry

Alternative is to push to the own doing the PR, but they would be able
to push another image to have the risk that in theory we test another PR
than was used in the code from the PR.

It seems to try to push the latest tag so make which tag pushed
explicit.