Remove apt lists after 'apt-get update' in docker-contributor/Dockerfile
We already do this in all other places. It reduces the size of the
image.

(Note that the 'rm' command has to be done in the same RUN instruction
as 'apt-get update', otherwise the lists will still be stored in the
intermediate layers. So some of the RUN instructions had to be
merged.)