Upgrade gitlab image to 24.04
Also removed some unneeded packages as we don't run those tests anymore.
We need some extra changes for the pa11y now as the chrome-browser gets
installed in an unreachable location for the domjudge user.
Also download the latest dictionary file now as we use the system
codespell now.
Composer is now again in the ubuntu repo's so we reverted to using that
one.