Move php-compatibility maintainance to our own container
For [phpcs_compatibility](https://github.com/DOMjudge/domjudge/blob/main/.github/workflows/codestyle.yml).

We used another container for simplicity of maintainance, then we needed to hotpatch the container for PHP84 for the `$this` & `enums`. The plan was to do this while waiting for version 10 of the tool. Since then we've had to patch the container more and more so the "lower" maintenance cost are now in installing the tool on it's own.